Xilsecure:Use of mixed mode arithmetic.
This patch fixes below MISRA-C violation:
Use of mixed mode arithmetic.
MISRA-C:2012/AMD1 R.10.1,R.10.2,R.10.3,R.10.4


Acked-for-series: VNSL Durga Challa <vnsl.durga.challa@xilinx.com>

+++ lib/sw_services/xilsecure/src/zynqmp/xsecure.c
@@ -1683,7 +1683,7 @@ static inline u32 XSecure_DataAuthentication(XSecure_AuthParam *AuthParam)
 		goto END;
 	}
 
-	if (XSECURE_CSU_KECCAK_SHA3 == AuthParam->PaddingType) {
+	if ((u8)XSECURE_CSU_KECCAK_SHA3 == AuthParam->PaddingType) {
 		Status = (u32)XSecure_Sha3PadSelection(&SecureSha3,
 		                                      XSECURE_CSU_KECCAK_SHA3);
 		if (Status != (u32)XST_SUCCESS) {
